Refactor user queries as React hook
