+++ src/features/account/components/Account.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Account: React.FC = () => {
+  return <div />;
+};
+
+export default Account;
