+++ src/pages/admin.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Container } from '@material-ui/core';
 
 import Permissions from '~/features/admin/components/Permissions';
-import PleaseLogin from '~/features/auth/components/PleaseLogin';
+import PleaseLogin from '~/features/account/components/PleaseLogin';
 import { ADMIN_USER_ROLE } from '~/server/constants';
 
 const Admin: React.FC = () => {
