+++ src/pages/join.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Container } from '@material-ui/core';
 
-import Join from '~/features/auth/components/Join';
+import Join from '~/features/account/components/Join';
 
 const JoinPage: React.FC = () => {
   return (
