+++ src/pages/recover.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { Container } from '@material-ui/core';
 
-import PleaseLogin from '~/features/auth/components/PleaseLogin';
-import RecoverPassword from '~/features/auth/components/RecoverPassword';
+import PleaseLogin from '~/features/account/components/PleaseLogin';
+import RecoverPassword from '~/features/account/components/RecoverPassword';
 
 const RecoverPage: React.FC = () => {
   return (
