+++ src/pages/reset.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { Container } from '@material-ui/core';
 
-import PleaseLogin from '~/features/auth/components/PleaseLogin';
-import ResetPassword from '~/features/auth/components/ResetPassword';
+import PleaseLogin from '~/features/account/components/PleaseLogin';
+import ResetPassword from '~/features/account/components/ResetPassword';
 
 const RecoverPage: React.FC = () => {
   return (
